Move sanitize to emoji-data, cleanup
